+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [3.7] - 05-03-2018
 ### Added
 - Digests now include information about recent malware scans, uptime, latency, and backup stats 
 - Invoice creation is now available via [InvoiceNinja](https://www.invoiceninja.org/)'s API using `deploy --invoice`
@@ -204,8 +204,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Old monolithic script rewritten
 
 
-[Unreleased]: https://github.com/EMRL/deploy/compare/v3.6.7...HEAD
-[3.6.8]: https://github.com/EMRL/deploy/compare/v3.6.7...v3.6.8
+[Unreleased]: https://github.com/EMRL/deploy/compare/v3.7...HEAD
+[3.6.8]: https://github.com/EMRL/deploy/compare/v3.6.7...v3.7
 [3.6.7]: https://github.com/EMRL/deploy/compare/v3.6.6...v3.6.7
 [3.6.6]: https://github.com/EMRL/deploy/compare/v3.6.5...v3.6.6
 [3.6.5]: https://github.com/EMRL/deploy/compare/v3.6.4...v3.6.5